Create a vector library
Starting the process to create a library that will revolve around
the use of vector structs that I will use in place of arrays.
Created a generic vector struct that has a void* array and the
variable size to indicate the size of each element in the array.
Also created a function for intializing the vectors in programs.